Rename module
SDESK-7287
